feat: impedir exclusão de contêiner com aluguel ativo

+++ polaris/src/app/gerente/visualizar-estoque-conteineres/visualizar-estoque-conteineres.component.ts
@@ -1,5 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { GerenteService } from '../services';
+import { HttpResponse } from '@angular/common/http';
+import { AluguelEstado } from 'src/app/models/aluguelEstado.model';
+import { Conteiner } from 'src/app/models/conteiner.model';
+import { Aluguel } from 'src/app/models/aluguel.model';
 
 enum EstadoConteiner {
   Cancelado = 0,
@@ -20,6 +24,8 @@ enum EstadoConteiner {
 })
 export class VisualizarEstoqueConteineresComponent implements OnInit {
   public conteinerData: any;
+  public estadoConteiner: any;
+  public statusMsg!: string;
   
   constructor( private gerenteService: GerenteService ) { }
 
@@ -42,10 +48,31 @@ export class VisualizarEstoqueConteineresComponent implements OnInit {
     }
   };
 
-  didTapOnDelete(uuid: string) {
-    this.gerenteService.deleteConteiner(uuid).subscribe(
-      response => this.loadData(),
-      error => console.error(error)
-    )
-  };
+  didTapOnDelete(uuid: string, codigo: string) {
+    this.gerenteService.getAlugueisByConteiner(codigo).subscribe(      
+      (res: any) => {
+        const estadoAluguelValues = res.map((item: any) => item.estadoAluguel);
+        
+        if (estadoAluguelValues.includes(0) || estadoAluguelValues.includes(1) || estadoAluguelValues.includes(2) || estadoAluguelValues.includes(3) || estadoAluguelValues.includes(5)) {
+          this.statusMsg = 'fail';
+        } else {
+          this.gerenteService.deleteConteiner(uuid).subscribe(
+            (response: HttpResponse<Conteiner>) => {   
+              if (response.status === 200 || response.status === 201) {
+                this.statusMsg = 'success';
+                this.loadData();
+                console.log('Put request successful');
+              } else {
+                console.log('Put request failed');
+              }
+            },
+            (error) => {
+              console.error('Error:', error);
+              this.statusMsg = 'fail';
+            }
+          );
+        }
+      }
+    );
+  }
 }
